+++ mlprogram/datasets/__init__.py
@@ -1,4 +1,5 @@
-import os
-
-DEFAULT_CACHE_DIR = \
-    os.path.join(os.environ["HOME"], ".cache", "mlprogram", "datasets")
+from mlprogram.datasets import deepfix  # noqa
+from mlprogram.datasets import django  # noqa
+from mlprogram.datasets import hearthstone  # noqa
+from mlprogram.datasets import nl2bash  # noqa
+from mlprogram.datasets.constants import DEFAULT_CACHE_DIR  # noqa
