+++ mlprogram/datasets/constants.py
@@ -1,0 +1,3 @@
+import os
+
+DEFAULT_CACHE_DIR = os.path.join(os.environ["HOME"], ".cache", "mlprogram", "datasets")
