+++ mlprogram/entrypoint/modules/__init__.py
@@ -1,0 +1,4 @@
+import mlprogram.entrypoint.modules.fairseq as fairseq  # noqa
+import mlprogram.entrypoint.modules.numpy as numpy  # noqa
+import mlprogram.entrypoint.modules.torch as torch  # noqa
+import mlprogram.entrypoint.modules.torchnlp as torchnlp  # noqa
