+++ mlprogram/entrypoint/modules/fairseq.py
@@ -1,5 +1,9 @@
-import fairseq.optim
+try:
+    import fairseq.optim
 
-types = {
-    "fairseq.optim.Adafactor": lambda: fairseq.optim.adafactor.Adafactor
-}
+    types = {
+        "fairseq.optim.Adafactor": lambda: fairseq.optim.adafactor.Adafactor
+    }
+except:  # noqa
+    types = {}
+    pass
